Changing int64_t to size_t. Keep the size_t used accross all monitoring components. Adapt the documentation. Remove useless MPI_Request and MPI_Status from monitoring_test.c.

+++ ompi/mca/common/monitoring/common_monitoring.h
@@ -111,9 +111,9 @@ OMPI_DECLSPEC OBJ_CLASS_DECLARATION(mca_monitoring_coll_data_t);
 
 OMPI_DECLSPEC mca_monitoring_coll_data_t*mca_common_monitoring_coll_new(ompi_communicator_t*comm);
 OMPI_DECLSPEC void mca_common_monitoring_coll_release(mca_monitoring_coll_data_t*data);
-OMPI_DECLSPEC void mca_common_monitoring_coll_o2a(uint64_t size, mca_monitoring_coll_data_t*data);
-OMPI_DECLSPEC void mca_common_monitoring_coll_a2o(uint64_t size, mca_monitoring_coll_data_t*data);
-OMPI_DECLSPEC void mca_common_monitoring_coll_a2a(uint64_t size, mca_monitoring_coll_data_t*data);
+OMPI_DECLSPEC void mca_common_monitoring_coll_o2a(size_t size, mca_monitoring_coll_data_t*data);
+OMPI_DECLSPEC void mca_common_monitoring_coll_a2o(size_t size, mca_monitoring_coll_data_t*data);
+OMPI_DECLSPEC void mca_common_monitoring_coll_a2a(size_t size, mca_monitoring_coll_data_t*data);
 
 END_C_DECLS
